+++ Teapot/Localizable.strings
@@ -1,3 +1,3 @@
-"Teapot:InvalidRequestPath" = "An error occurred: request URL path is invalid.";
-"Teapot:MissingImage" = "An error occurred: image is missing.";
-"Teapot:InvalidResponseStatus" = "An error occurred: request response status reported an issue. Status code: %d.";
+"teapot_invalid_request_path" = "An error occurred: request URL path is invalid.";
+"teapot_missing_image" = "An error occurred: image is missing.";
+"teapot_invalid_response_status" = "An error occurred: request response status reported an issue. Status code: %d.";
