+++ Teapot/Teapot.swift
@@ -2,28 +2,7 @@ import Foundation
 
 /// A light-weight abstraction for URLSession.
 open class Teapot {
-    static var localizationBundle = Bundle.init(for: Teapot.self)
-
-    public struct TeapotError: LocalizedError {
-        static let invalidRequestPath = TeapotError(responseStatus: -1, type: .invalidRequestPath, errorDescription: NSLocalizedString("Teapot:InvalidRequestPath", bundle: Teapot.localizationBundle, comment: ""))
-
-        static let missingImage = TeapotError(responseStatus: -1, type: .missingImage, errorDescription: NSLocalizedString("Teapot:MissingImage", bundle: Teapot.localizationBundle, comment: ""))
-
-        static func invalidResponseStatus(_ status: Int) -> TeapotError {
-            return TeapotError(responseStatus: status, type: .invalidResponseStatus, errorDescription: String(format: NSLocalizedString("Teapot:InvalidResponseStatus", bundle: Teapot.localizationBundle, comment: ""), status))
-        }
-
-        enum ErrorType {
-            case invalidRequestPath
-            case invalidResponseStatus
-            case missingImage
-        }
-
-        let responseStatus: Int
-        let type: ErrorType
-
-        public var errorDescription: String?
-    }
+    open static var localizationBundle = Bundle.init(for: Teapot.self)
 
     /// The URL request verb to be passed to the URLRequest.
     enum Verb: String {
